+++ packages/desktop/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @veltra/desktop
 
+## 1.0.7
+
+### Patch Changes
+
+- a91e7a8: 新增面包屑组件
+- Updated dependencies [a91e7a8]
+  - @veltra/compositions@1.0.7
+  - @veltra/directives@1.0.7
+  - @veltra/icons@1.0.7
+  - @veltra/styles@1.0.7
+  - @veltra/utils@1.0.7
+
 ## 1.0.6
 
 ### Patch Changes