+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @veltra/icons
 
+## 1.0.7
+
+### Patch Changes
+
+- a91e7a8: 新增面包屑组件
+
 ## 1.0.6
 
 ### Patch Changes